+++ test/clippy/README.txt
@@ -1,3 +1,3 @@
 1) make a venv with the pacakges in requirements.txt (and optionally requirements-dev.txt)
 2) activate the venv
-3) CLIPPY_BACKEND_PATH=/workspaces/metalldata/build/src/clippy pytest -v .
+3) CLIPPY_BACKEND_PATH=/workspaces/metalldata/build/src/clippy DATA_DIR=../../data/metall_graph pytest tests/
