+++ test/clippy/pytest.ini
@@ -1,5 +1,2 @@
 [pytest]
-markers =
-    order: execution order for tests
-
 tmp_path_retention_policy = none
