+++ test/clippy/requirements.txt
@@ -1,4 +1,3 @@
 pytest
 llnl-clippy
 pytest-cov
-pytest-ordering
