--- test/clippy/test_mg.py
@@ -1,509 +1,0 @@
-from typing import Any
-import pytest
-import clippy
-from clippy import MetallGraph # type: ignore
-import os
-from clippy.error import ClippyValidationError, ClippyInvalidSelectorError
-from clippy.backends.fs.execution import NonZeroReturnCodeError
-import logging
-import shutil
-
-DATA_DIR = os.environ.get("DATA_DIR", "../../data/metall_graph")
-
-clippy.logger.setLevel(logging.WARN)
-logging.getLogger().setLevel(logging.WARN)
-
-def is_as_described(mg: MetallGraph, nv: int, ne: int, check_for_path: bool = True):
-    r = mg.describe()
-    assert "nv" in r
-    assert "ne" in r
-    if check_for_path:
-        assert "path" in r
-    assert r["nv"] == nv
-    assert r["ne"] == ne
-
-def is_as_selected(data_list: list[dict], required_pairs: dict, required_fields: list[str], missing_fields: list[str]):
-    for el in data_list:
-        assert required_pairs.items() <= el.items()
-        for missing in missing_fields:
-            assert missing not in el.keys()
-        for required in required_fields:
-            assert required in el.keys()
-
-def is_specific(data_list: list[dict[str, Any]], key: str, keydict: dict[str, dict[str, Any]]):
-    for el in data_list:
-        if key in el:  # "id"
-            keyval = el[key] # "path-a" or similar node value
-            if keyval in keydict:
-                assert keydict[keyval].items() <= el.items() and len(keydict[keyval].items()) > 0
-            
-                
-
-
-    
-@pytest.fixture(scope="module")
-def temp_file(tmp_path_factory):
-    temp_name = tmp_path_factory.mktemp("data") / "metallgraph.db"
-    shutil.rmtree(str(temp_name), ignore_errors=True)
-    yield str(temp_name)
-
-
-@pytest.fixture()
-def metallgraph(temp_file):
-    return MetallGraph(temp_file)
-
-@pytest.mark.order(1)
-def test_mg_create(metallgraph):
-    is_as_described(metallgraph, 0, 0)
-
-@pytest.mark.order(2)
-def test_mg_ingest_parquet(metallgraph):
-    metallgraph.ingest_parquet_edges(DATA_DIR + "/test", "s", "t")
-    is_as_described(metallgraph, 21, 28)
-    el = metallgraph.select_edges()
-    is_as_selected(el, {}, ["u", "v", "graphnum", "relevant"], ["foo", "field_does_not_exist"])
-
-@pytest.mark.order(3)
-def test_mg_assign(metallgraph):
-    metallgraph.assign("node.assign1", True)
-    select_data = metallgraph.select_nodes()
-    is_as_selected(select_data, {}, ["id", "assign1"], ["field_does_not_exist"])
-    metallgraph.assign("edge.assign1", True)
-    select_data = metallgraph.select_edges()
-    is_as_selected(select_data, {"assign1": True}, ["u", "v"], ["field_does_not_exist"])
-
-    metallgraph.assign("edge.assign2", True, where=metallgraph.edge.graphnum == 0)
-    select_data = metallgraph.select_edges(where=metallgraph.edge.graphnum == 0)
-    is_as_selected(select_data, {"assign2": True}, ["u", "v"], ["field_does_not_exist"])
-    select_data = metallgraph.select_edges(where=metallgraph.edge.graphnum != 0)
-    is_as_selected(select_data, {}, ["u", "v"], ["field_does_not_exist", "assign2"])
-
-    # assign node based on edge where
-    # used in test nhops
-    metallgraph.assign("node.gnum", 3, where=metallgraph.edge.graphnum == 3)
-    select_data = metallgraph.select_nodes(where=metallgraph.edge.graphnum == 3)
-    is_as_selected(select_data, {"gnum": 3}, [], [])
-    select_data = metallgraph.select_nodes(where=metallgraph.edge.graphnum != 3)
-    is_as_selected(select_data, {}, [], ["gnum"])
-
-
-@pytest.mark.order(4)
-def test_mg_add_faker(metallgraph):
-    metallgraph.add_faker("node.randpct", "percentage")
-    select_data = metallgraph.select_nodes()
-    is_as_selected(select_data, {}, ["randpct", "id"], ["field_does_not_exist"])
-
-    metallgraph.add_faker("node.uuid", "uuid4")
-    select_data = metallgraph.select_nodes()
-    is_as_selected(select_data, {}, ["uuid", "id"], ["field_does_not_exist"])
-    metallgraph.add_faker("edge.temp", "double", where=metallgraph.edge.graphnum == 0)
-    select_data = metallgraph.select_edges(where=metallgraph.edge.graphnum == 0)
-    is_as_selected(select_data, {}, ["u", "v", "temp"], ["field_does_not_exist"])
-    select_data = metallgraph.select_edges(where=metallgraph.edge.graphnum != 0)
-    is_as_selected(select_data, {}, ["u", "v"], ["temp"])
-
-
-@pytest.mark.order(5)
-def test_mg_drop_series(metallgraph):
-
-    select_data = metallgraph.select_nodes()
-    is_as_selected(select_data, {}, ["id", "assign1"], ["field_does_not_exist"])
-    metallgraph.drop_series(metallgraph.node.assign1)
-    select_data = metallgraph.select_nodes()
-    is_as_selected(select_data, {}, ["id"], ["assign1", "field_does_not_exist"])
-
-    select_data = metallgraph.select_edges()
-    is_as_selected(select_data, {}, ["u", "assign1"], ["field_does_not_exist"])
-    metallgraph.drop_series(metallgraph.edge.assign1)
-    select_data = metallgraph.select_edges()
-    is_as_selected(select_data, {}, ["u"], ["assign1", "field_does_not_exist"])
-    
-
-@pytest.mark.order(6)
-def test_mg_nhops(metallgraph):    
-    metallgraph.nhops("nhops", 2, ["path-a", "path-f"])
-    select_data = metallgraph.select_nodes()
-    required_result = {"path-a": {"nhops": 0},
-                       "path-f": {"nhops": 0},
-                       "path-b": {"nhops": 1},
-                       "path-g": {"nhops": 1},
-                       "path-c": {"nhops": 2}
-                       }
-    
-    for d in select_data:
-        if "nhops" in d:
-            assert d["nhops"] <= 2
-
-    is_specific(select_data, "id", required_result)
-    select_data = metallgraph.select_nodes(where=metallgraph.node.gnum != 3)
-    is_as_selected(select_data, {}, ["id"], ["nhops"])
-
-
-@pytest.mark.order(7)
-def test_mg_topk(metallgraph):    
-
-    # test edges with k = 5 and additional data
-    results = metallgraph.topk(metallgraph.edge.randint, k=5, addl_series=[metallgraph.edge.name])
-
-    assert (len(results) == 5)
-    curr_max = 10000;
-    for result in results:
-        assert len(result) == 2
-        assert result[0] <= curr_max
-        curr_max = result[0]
-
-    # test nodes with k > numrows
-    results = metallgraph.topk(metallgraph.node.randpct, k=50000)
-
-    assert (len(results) == metallgraph.describe()["nv"])
-
-    curr_max = 10000;
-    for result in results:
-        assert len(result) == 1
-        assert result[0] <= curr_max
-        curr_max = result[0]
-
-    # test edges with where clause
-    results = metallgraph.topk(metallgraph.edge.randint, where=metallgraph.edge.graphnum == 1)
-    rresults = [x[0] for x in results]
-    assert (rresults == [92, 84, 76, 57, 33, 19])
-
-
-@pytest.mark.order(8)
-def test_mg_select_sample_edges(metallgraph):
-    # test basic sampling with k < total edges
-    for _ in range(20):
-        sample_data = metallgraph.select_sample_edges(k=10, series_names=[metallgraph.edge.u, metallgraph.edge.v])
-        assert len(sample_data) == 10
-        is_as_selected(sample_data, {}, ["u", "v"], [])
-
-    # test sampling with where clause
-    for _ in range(5):
-        sample_data = metallgraph.select_sample_edges(k=5, series_names=[metallgraph.edge.u, metallgraph.edge.v, metallgraph.edge.graphnum], where=metallgraph.edge.graphnum == 0)
-        assert len(sample_data) == 5
-        for el in sample_data:
-            assert el["graphnum"] == 0
-        is_as_selected(sample_data, {"graphnum": 0}, ["u", "v"], [])
-
-    # test sampling with k > available edges with where clause
-    for _ in range(5):
-        sample_data = metallgraph.select_sample_edges(k=1000, series_names=[metallgraph.edge.u, metallgraph.edge.v, metallgraph.edge.graphnum], where=metallgraph.edge.graphnum == 1)
-        total_edges_g1 = len(metallgraph.select_edges(where=metallgraph.edge.graphnum == 1))
-        assert len(sample_data) <= total_edges_g1
-        is_as_selected(sample_data, {"graphnum": 1}, ["u", "v"], [])
-
-    # test sampling all edges when k >= total edges
-    for _ in range(5):
-        total_edges = metallgraph.describe()["ne"]
-        sample_data = metallgraph.select_sample_edges(k=total_edges + 10)
-        assert len(sample_data) <= total_edges
-
-    # test seeds
-    for seed in range(5):
-        sd1 = metallgraph.select_sample_edges(10, series_names=[metallgraph.edge.u, metallgraph.edge.v, metallgraph.edge.graphnum], seed=seed)
-
-        sd2 = metallgraph.select_sample_edges(10, series_names=[metallgraph.edge.u, metallgraph.edge.v, metallgraph.edge.graphnum], seed=seed)
-
-        sd3 = metallgraph.select_sample_edges(10, [metallgraph.edge.u, metallgraph.edge.v, metallgraph.edge.graphnum], seed=seed+1)
-
-        assert len(sd1) == len(sd2) == len(sd3)
-        assert sd1 == sd2
-        assert sd2 != sd3
-
-
-@pytest.mark.order(9)
-def test_mg_sample_edges(metallgraph):
-    # test column creation
-    metallgraph.sample_edges(k=10, series_name="samp1")
-    select_data = metallgraph.select_edges(where=metallgraph.edge.samp1 == True)
-    is_as_selected(select_data, {"samp1": True}, ["u", "v"], [])
-    assert len(select_data) == 10
-
-    metallgraph.sample_edges(k=3, series_name="samp2", where=metallgraph.edge.graphnum == 1)
-    select_data = metallgraph.select_edges(where=metallgraph.edge.samp2 == True)
-    is_as_selected(select_data, {"samp2": True, "graphnum": 1}, ["u", "v"], [])
-    assert len(select_data) == 3
-
-
-@pytest.mark.order(10)
-def test_mg_select_sample_nodes(metallgraph):
-    for _ in range(20):
-        sample_data = metallgraph.select_sample_nodes(
-            k=10, series_names=[metallgraph.node.id]
-        )
-        assert len(sample_data) == 10
-        is_as_selected(sample_data, {}, ["id"], [])
-
-    for _ in range(5):
-        sample_data = metallgraph.select_sample_nodes(
-            k=5,
-            series_names=[metallgraph.node.id, metallgraph.node.gnum],
-            where=metallgraph.node.gnum == 3,
-        )
-        assert len(sample_data) == 5
-        for el in sample_data:
-            assert el["gnum"] == 3
-        is_as_selected(sample_data, {"gnum": 3}, ["id"], [])
-
-    for _ in range(5):
-        sample_data = metallgraph.select_sample_nodes(
-            k=1000,
-            series_names=[metallgraph.node.id, metallgraph.node.gnum],
-            where=metallgraph.node.gnum == 3,
-        )
-        total_nodes_g3 = len(
-            metallgraph.select_nodes(where=metallgraph.node.gnum == 3)
-        )
-        assert len(sample_data) <= total_nodes_g3
-        is_as_selected(sample_data, {"gnum": 3}, ["id"], [])
-
-    for _ in range(5):
-        total_nodes = metallgraph.describe()["nv"]
-        sample_data = metallgraph.select_sample_nodes(k=total_nodes + 10)
-        assert len(sample_data) <= total_nodes
-
-    for seed in range(5):
-        sd1 = metallgraph.select_sample_nodes(
-            10, series_names=[metallgraph.node.id, metallgraph.node.gnum], seed=seed
-        )
-        sd2 = metallgraph.select_sample_nodes(
-            10, series_names=[metallgraph.node.id, metallgraph.node.gnum], seed=seed
-        )
-        sd3 = metallgraph.select_sample_nodes(
-            10,
-            series_names=[metallgraph.node.id, metallgraph.node.gnum],
-            seed=seed + 1,
-        )
-
-        assert len(sd1) == len(sd2) == len(sd3)
-        assert sd1 == sd2
-        assert sd2 != sd3
-
-
-@pytest.mark.order(11)
-def test_mg_sample_nodes(metallgraph):
-    metallgraph.sample_nodes(k=10, series_name="nsamp1")
-    select_data = metallgraph.select_nodes(where=metallgraph.node.nsamp1 == True)
-    is_as_selected(select_data, {"nsamp1": True}, ["id"], [])
-    assert len(select_data) == 10
-
-    metallgraph.sample_nodes(
-        k=3, series_name="nsamp2", where=metallgraph.node.gnum == 3
-    )
-    select_data = metallgraph.select_nodes(where=metallgraph.node.nsamp2 == True)
-    is_as_selected(select_data, {"nsamp2": True, "gnum": 3}, ["id"], [])
-    assert len(select_data) == 3
-
-
-@pytest.mark.order(12)
-def test_mg_rename_series(metallgraph):
-    metallgraph.rename_series(metallgraph.node.nsamp1, "testrename1")
-    select_data = metallgraph.select_nodes(where=metallgraph.node.testrename1 == True)
-    is_as_selected(select_data, {"testrename1": True}, ["id"], [])
-
-    metallgraph.rename_series(metallgraph.node.testrename1, "node.testrename2")
-    select_data = metallgraph.select_nodes(where=metallgraph.node.testrename2 == True)
-    is_as_selected(select_data, {"testrename2": True}, ["id"], [])
-
-    metallgraph.rename_series(metallgraph.node.testrename2, "nsamp1")
-    select_data = metallgraph.select_nodes(where=metallgraph.node.nsamp1 == True)
-    is_as_selected(select_data, {"nsamp1": True}, ["id"], [])
-
-    metallgraph.rename_series(metallgraph.edge.samp1, "testrename1")
-    select_data = metallgraph.select_edges(where=metallgraph.edge.testrename1 == True)
-    is_as_selected(select_data, {"testrename1": True}, ["u", "v"], [])
-
-    metallgraph.rename_series(metallgraph.edge.testrename1, "edge.testrename2")
-    select_data = metallgraph.select_edges(where=metallgraph.edge.testrename2 == True)
-    is_as_selected(select_data, {"testrename2": True}, ["u", "v"], [])
-
-    metallgraph.rename_series(metallgraph.edge.testrename2, "samp1")
-    select_data = metallgraph.select_edges(where=metallgraph.edge.samp1 == True)
-    is_as_selected(select_data, {"samp1": True}, ["u", "v"], [])
-
-    with pytest.raises(NonZeroReturnCodeError):
-        metallgraph.rename_series(metallgraph.edge.samp1, "edge.u")
-        metallgraph.rename_series(metallgraph.node.nsamp1, "node.id")
-
-        metallgraph.rename_series(metallgraph.edge.samp1, "node.testrename")
-        metallgraph.rename_series(metallgraph.node.nsamp1, "edge.testrename")
-        metallgraph.rename_series(metallgraph.edge.samp1, "samp1")
-        metallgraph.rename_series(metallgraph.edge.samp1, "edge.samp1")
-        metallgraph.rename_series(metallgraph.node.nsamp1, "nsamp1")
-        metallgraph.rename_series(metallgraph.node.nsamp1, "node.nsamp1")
-
-
-@pytest.mark.order(12)
-def test_mg_indegree(metallgraph):
-    degs1 = {
-        "5clique-a": {"indeg1": 0},
-        "5clique-b": {"indeg1": 1},
-        "5clique-c": {"indeg1": 2},
-        "5clique-d": {"indeg1": 3},
-        "5clique-e": {"indeg1": 4}
-    }
-
-    degs2 = {
-        "5clique-d": {"indeg1": 3},
-        "5clique-e": {"indeg1": 4}
-    }
-
-
-    metallgraph.in_degree("indeg1")
-    select_data = metallgraph.select_nodes()
-    is_specific(select_data, "id", degs1)
-
-    metallgraph.in_degree("indeg2", where = metallgraph.node.indeg1 > 2)
-    select_data = metallgraph.select_nodes()
-    is_specific(select_data, "id", degs2)
-
-
-@pytest.mark.order(12)
-def test_mg_outdegree(metallgraph):
-    degs1 = {
-        "5clique-a": {"outdeg1": 4},
-        "5clique-b": {"outdeg1": 3},
-        "5clique-c": {"outdeg1": 2},
-        "5clique-d": {"outdeg1": 1},
-        "5clique-e": {"outdeg1": 0}
-    }
-
-    degs2 = {
-        "5clique-a": {"outdeg1": 4},
-        "5clique-b": {"outdeg1": 3}
-    }
-
-
-    metallgraph.out_degree("outdeg1")
-    select_data = metallgraph.select_nodes()
-    is_specific(select_data, "id", degs1)
-
-    metallgraph.out_degree("outdeg2", where = metallgraph.node.outdeg1 > 2)
-    select_data = metallgraph.select_nodes()
-    is_specific(select_data, "id", degs2)
-
-
-# {'id': '5clique-a', 'indeg1': 0, 'randpct': 4.519041400698163, 'uuid': 'bffdb73b-e003-4904-b4c4-fcfb6de44ded'}, {'id': '5clique-b', 'indeg1': 1, 'nsamp1': True, 'randpct': 18.729939416659157, ...}, {'id': '5clique-c', 'indeg1': 2, 'randpct': 57.53399788992132, 'uuid': 'bffdb73b-e003-4904-b4c4-fcfb6de44ded'}, {'id': '5clique-d', 'indeg1': 3, 'randpct': 3.4017999660346714, 'uuid': 'bffdb73b-e003-4904-b4c4-fcfb6de44ded'}
-
-@pytest.mark.order(13)
-def test_mg_nunique_edge(metallgraph):
-    # int series: graphnum takes values 0,1,2,3 across all subgraphs
-    result = metallgraph.nunique(series_names=["edge.graphnum"])
-    assert result["edge.graphnum"] == 4
-
-    # string series: 8 distinct colors across all edges (nulls excluded)
-    result = metallgraph.nunique(series_names=["edge.color"])
-    assert result["edge.color"] == 8
-
-    # bool series: relevant is True in path_graph/four_clique, False in five_clique/two_triangles
-    result = metallgraph.nunique(series_names=["edge.relevant"])
-    assert result["edge.relevant"] == 2
-
-    # bool series with only one value: samp1 marks exactly k edges True, rest unset
-    result = metallgraph.nunique(series_names=["edge.samp1"])
-    assert result["edge.samp1"] == 1
-
-    # int series with many values
-    result = metallgraph.nunique(series_names=["edge.randint"])
-    assert result["edge.randint"] == 22
-
-    # multiple series in one call
-    result = metallgraph.nunique(series_names=["edge.graphnum", "edge.color", "edge.relevant"])
-    assert result["edge.graphnum"] == 4
-    assert result["edge.color"] == 8
-    assert result["edge.relevant"] == 2
-
-
-@pytest.mark.order(13)
-def test_mg_nunique_edge_where(metallgraph):
-    # graphnum==0 edges are only from two_triangles — one distinct graphnum value
-    result = metallgraph.nunique(series_names=["edge.graphnum"], where=metallgraph.edge.graphnum == 0)
-    assert result["edge.graphnum"] == 1
-
-    # two_triangles has only "blue" and "red" edges
-    result = metallgraph.nunique(series_names=["edge.color"], where=metallgraph.edge.graphnum == 0)
-    assert result["edge.color"] == 2
-
-    # two_triangles has only False for relevant
-    result = metallgraph.nunique(series_names=["edge.relevant"], where=metallgraph.edge.graphnum == 0)
-    assert result["edge.relevant"] == 1
-
-
-@pytest.mark.order(13)
-def test_mg_nunique_node(metallgraph):
-    # all 21 node IDs are distinct
-    result = metallgraph.nunique(series_names=["node.id"])
-    assert result["node.id"] == 21
-    # gnum was assigned value 3 for all nodes reachable via graphnum==3 edges — only one distinct value
-    result = metallgraph.nunique(series_names=["node.gnum"])
-    assert result["node.gnum"] == 1
-
-
-@pytest.mark.order(13)
-def test_mg_nunique_mixed_where(metallgraph):
-    # path_graph is graphnum==3: 6 edges, 7 nodes (path-a through path-g)
-    result = metallgraph.nunique(
-        series_names=["edge.graphnum", "edge.color", "node.id", "node.gnum"],
-        where=metallgraph.edge.graphnum == 3
-    )
-    assert result["edge.graphnum"] == 1   # only value 3
-    assert result["edge.color"] == 6      # red, orange, yellow, green, blue, indigo
-    assert result["node.id"] == 7         # path-a through path-g
-    assert result["node.gnum"] == 1       # only value 3
-
-
-@pytest.mark.order(14)
-def test_mg_value_counts_edge(metallgraph):
-    # edge.graphnum has 4 distinct values; total edges = 28
-    results = metallgraph.value_counts(metallgraph.edge.graphnum, k=100)
-    assert len(results) == 4
-    assert sum(r[1] for r in results) == 28
-    # five_clique (graphnum==2) has 10 edges — the most
-    counts = {r[0]: r[1] for r in results}
-    assert counts[2] == 10
-
-
-@pytest.mark.order(14)
-def test_mg_value_counts_top_bottom_k(metallgraph):
-    results_top = metallgraph.value_counts(metallgraph.edge.graphnum, k=2)
-    assert len(results_top) == 2
-
-    results_bottom = metallgraph.value_counts(metallgraph.edge.graphnum, k=-2)
-    assert len(results_bottom) == 2
-
-    # top and bottom values should not overlap (five_clique is uniquely the largest)
-    top_vals = {r[0] for r in results_top}
-    bottom_vals = {r[0] for r in results_bottom}
-    assert 2 in top_vals
-    assert 2 not in bottom_vals
-
-
-@pytest.mark.order(14)
-def test_mg_value_counts_where(metallgraph):
-    # restrict to graphnum==0 edges — only one distinct value
-    results = metallgraph.value_counts(
-        metallgraph.edge.graphnum, k=100, where=metallgraph.edge.graphnum == 0
-    )
-    assert len(results) == 1
-    assert results[0][0] == 0
-    assert results[0][1] == 6  # two_triangles has 6 edges
-
-
-@pytest.mark.order(14)
-def test_mg_value_counts_bool(metallgraph):
-    # relevant is True for four_clique+path_graph (12 edges), False for five_clique+two_triangles (16)
-    results = metallgraph.value_counts(metallgraph.edge.relevant, k=100)
-    assert len(results) == 2
-    counts = {r[0]: r[1] for r in results}
-    assert counts[True] == 12
-    assert counts[False] == 16
-
-
-@pytest.mark.order(14)
-def test_mg_value_counts_node(metallgraph):
-    # node.gnum was assigned value 3 to nodes reachable via graphnum==3 edges;
-    # remaining 14 nodes have no value (None/monostate)
-    results = metallgraph.value_counts(metallgraph.node.gnum, k=100)
-    counts = {r[0]: r[1] for r in results}
-    assert counts[3] == 7
-    assert counts[None] == 14
