+++ test/clippy/tests/test_mg_select_sample_nodes.py
@@ -1,0 +1,51 @@
+from conftest import is_as_selected
+
+
+def test_mg_select_sample_nodes(metallgraph):
+    for _ in range(20):
+        sample_data = metallgraph.select_sample_nodes(
+            k=10, series_names=[metallgraph.node.id]
+        )
+        assert len(sample_data) == 10
+        is_as_selected(sample_data, {}, ["id"], [])
+
+    for _ in range(5):
+        sample_data = metallgraph.select_sample_nodes(
+            k=5,
+            series_names=[metallgraph.node.id, metallgraph.node.gnum],
+            where=metallgraph.node.gnum == 3,
+        )
+        assert len(sample_data) == 5
+        for el in sample_data:
+            assert el["gnum"] == 3
+        is_as_selected(sample_data, {"gnum": 3}, ["id"], [])
+
+    for _ in range(5):
+        sample_data = metallgraph.select_sample_nodes(
+            k=1000,
+            series_names=[metallgraph.node.id, metallgraph.node.gnum],
+            where=metallgraph.node.gnum == 3,
+        )
+        total_nodes_g3 = len(metallgraph.select_nodes(where=metallgraph.node.gnum == 3))
+        assert len(sample_data) <= total_nodes_g3
+        is_as_selected(sample_data, {"gnum": 3}, ["id"], [])
+
+    for _ in range(5):
+        total_nodes = metallgraph.describe()["nv"]
+        sample_data = metallgraph.select_sample_nodes(k=total_nodes + 10)
+        assert len(sample_data) <= total_nodes
+
+    for seed in range(5):
+        sd1 = metallgraph.select_sample_nodes(
+            10, series_names=[metallgraph.node.id, metallgraph.node.gnum], seed=seed
+        )
+        sd2 = metallgraph.select_sample_nodes(
+            10, series_names=[metallgraph.node.id, metallgraph.node.gnum], seed=seed
+        )
+        sd3 = metallgraph.select_sample_nodes(
+            10, series_names=[metallgraph.node.id, metallgraph.node.gnum], seed=seed + 1
+        )
+
+        assert len(sd1) == len(sd2) == len(sd3)
+        assert sd1 == sd2
+        assert sd2 != sd3
